requirements.txt
Django
gunicorn
django-heroku
django-allauth
django-bootstrap-v5